requirements.txt
cfg_param_wrapper>=1.0.1
imagehash
imagesize
numpy
opencv-python
pillow
polars
pyarrow
python-dateutil
rich
tqdm
typer[all]
pywinstyles; sys.platform=='win32'